Move useUniswapMarketDetails() into web3-utils
